--- .env.example
@@ -1,7 +1,0 @@
-# Create local .env file with these fields
-# Included in .gitignore
-
-# New York Times API Key 
-# React Usage: process.env.REACT_APP_NYT_KEY
-
-REACT_APP_NYT_KEY = YOUR-API-KEY
